requirements.txt
# Core LLM & Orchestration
langchain>=0.1.0
langchain-community>=0.0.10
langchain-google-genai>=0.0.6
langchain-openai>=0.0.2
langchain-text-splitters>=0.0.1

# Databases
neo4j>=5.15.0
weaviate-client>=4.4.0

# API & UI
fastapi>=0.109.0
uvicorn>=0.27.0
streamlit>=1.31.0
pydantic>=2.6.0
pydantic-settings>=2.1.0

# MLOps & Observability
mlflow>=2.10.0
ragas>=0.1.0
prometheus_client>=0.19.0
psutil>=5.9.0

# Utilities
python-dotenv>=1.0.1
pandas>=2.2.0
numpy>=1.26.0
pypdf>=4.0.0
beautifulsoup4>=4.12.0
tqdm>=4.66.0
datasets>=2.18.0
